add storage2 server and client classes; define message types and error codes. pull remote into it's own library.
