+++ packages/library/lib-remote/Client.lua
@@ -2,22 +2,10 @@
 package.path = package.path .. ";/usr/lib/?.lua"
 require("class-lua.class")
 
-local pretty = require("cc.pretty")
-
-require("lib-storage2.remote.Remote")
-require("lib-storage2.remote.RemoteMessageType")
-require("lib-storage2.remote.CommandType")
+require("lib-remote.Remote")
+require("lib-remote.types.MessageType")
 
 local logger = require("lexicon-lib.lib-logging").getLogger("Client")
-
-
-local SERVER_ID_SETTING_NAME = "storage2-remote.server-id"
-
-settings.define(SERVER_ID_SETTING_NAME, {
-    description = "The ID of the storage2 server to connect to",
-    type = "number",
-    default = nil,
-})
 
 
 ---@class Client: Remote
@@ -36,6 +24,14 @@ end
 ---Find the server to connect to
 ---@return number?
 function Client:findServer()
+    local SERVER_ID_SETTING_NAME = self.protocol .. ".server-id"
+
+    settings.define(SERVER_ID_SETTING_NAME, {
+        description = "The ID of the " .. self.protocol .. " server to connect to",
+        type = "number",
+        default = nil,
+    })
+
     local serverId = settings.get(SERVER_ID_SETTING_NAME)
     if serverId then
         logger:debug("Using server ID from settings: %d", serverId)
@@ -66,7 +62,7 @@ end
 
 
 ---Send a command to the server, handling any responses
----@param commandType CommandType
+---@param commandType string
 ---@param sendData? table
 ---@return boolean, table?
 function Client:baseSendCommand(commandType, sendData)
@@ -100,73 +96,6 @@ function Client:baseSendCommand(commandType, sendData)
 
     logger:error("Unexpected response: %s", messageType)
     return false
-end
-
-
----Send a refresh request to the server
----@return boolean, nil
-function Client:refresh()
-    local res, _ = self:baseSendCommand(CommandType.REFRESH)
-
-    if res then
-        return true
-    end
-
-    return false
-end
-
-
----Get the input and output chest names from the server
----@return boolean, table?
-function Client:getChestNames()
-    return self:baseSendCommand(CommandType.DATA_IO_CHESTS)
-end
-
-
----Ping the server, printing how long it took
----@return boolean, table?
-function Client:ping()
-    local start = os.clock()
-    local res, data = self:baseSendCommand(CommandType.PING)
-    local duration = os.clock() - start
-
-    if res then
-        logger:info("Pong! Took %.2f seconds", duration)
-    end
-
-    return res, data
-end
-
-
----Pull items from the storage chests to an inventory
----@param outputChestName string The name of the inventory to pull to
----@param item string
----@param count number
----@param toSlot number?
----@return boolean, table?
-function Client:pull(outputChestName, item, count, toSlot)
-    local res, data = self:baseSendCommand(CommandType.PULL, {
-        item = item,
-        count = count,
-        invName = outputChestName,
-        toSlot = toSlot,
-    })
-
-    return res, data
-end
-
-
----Push items from an inventory to the storage chests
----@param inputChestName string The name of the inventory to push from
----@param slots? number[]
----@return boolean, table?
-function Client:push(inputChestName, slots)
-    local res, data = self:baseSendCommand(CommandType.PUSH, {
-        invName = inputChestName,
-        fromSlots = slots,
-    })
-
-    return res, data
 end
