+++ packages/library/lib-remote/types/MessageErrorCode.lua
@@ -1,0 +1,6 @@
+---@enum MessageErrorCode
+MessageErrorCode = {  ---@class _MessageErrorCode
+    INVALID_DATA_TYPE = "INVALID_DATA_TYPE",
+    UNKNOWN_COMMAND = "UNKNOWN_COMMAND",
+    UNKNOWN = "UNKNOWN",
+}
