+++ packages/library/lib-remote/types/MessageType.lua
@@ -1,0 +1,7 @@
+---@enum MessageType
+MessageType = { ---@class _MessageType
+    ACK = "ACK",
+    END = "END",
+    ERR = "ERR",
+    CMD = "CMD",
+}
