+++ packages/library/lib-remote/types/MessageTypeArgs.lua
@@ -1,22 +1,3 @@
----Message types
-
----@enum MessageType
-MessageType = { ---@class _MessageType
-    ACK = "ACK",
-    END = "END",
-    ERR = "ERR",
-    CMD = "CMD",
-}
-
-
----@enum MessageErrorCode
-MessageErrorCode = {  ---@class _MessageErrorCode
-    INVALID_DATA_TYPE = "INVALID_DATA_TYPE",
-    UNKNOWN_COMMAND = "UNKNOWN_COMMAND",
-    UNKNOWN = "UNKNOWN",
-}
-
-
 ---@class MessageDataArg
 ---@field name string
 ---@field type string|string[]
@@ -53,18 +34,3 @@ MessageTypeArgs = {
         },
     },
 }
-
----@alias MessageEndData table?
-
----@class MessageErrorData
----@field code MessageErrorCode
----@field message? string
-
----@class MessageCommandData
----@field type string
----@field data? table
-
----@alias MessageData MessageErrorData|MessageCommandData|MessageEndData
-
-
----@alias CommandHandler fun(clientId: number, data?: table): boolean, table?
