+++ packages/library/lib-storage2/remote/StorageServer.lua
@@ -4,29 +4,22 @@ require("class-lua.class")
 
 local chestHelpers = require("lib-storage2.chestHelpers")
 require("lib-storage2.Map")
-require("lib-storage2.remote.Remote")
-require("lib-storage2.remote.RemoteMessageType")
+require("lib-remote.server.Server")
 require("lib-storage2.remote.CommandType")
-
-local logger = require("lexicon-lib.lib-logging").getLogger("Server")
 
 
 -- Class definition
 
----@class Server: Remote
----@overload fun(): Server
-Server = Remote:extend()
+---@class StorageServer: Server
+---@overload fun(): StorageServer
+StorageServer = Server:extend()
 
-Server.filterCommands = {
-    [MessageType.ACK] = true,
-    [MessageType.END] = true,
-    [MessageType.ERR] = true,
-}
+StorageServer.protocol = "storage2-remote"
+
 
 ---Initialise a new storage2 server
-function Server:init()
-    Remote.init(self)
-    self.hostname = self.protocol .. "-" .. os.getComputerID()
+function StorageServer:init()
+    Server.init(self)
 
     self.commandHandlers = {
         [CommandType.REFRESH] = self.handleRefresh,
@@ -42,7 +35,7 @@ end
 
 ---Initialise the storage-related peripherals
 ---@return boolean
-function Server:initPeripherals()
+function StorageServer:initPeripherals()
     local inputChest = chestHelpers.getInputChest()
     if not inputChest then
         return false
@@ -67,42 +60,11 @@ function Server:initPeripherals()
 end
 
 
----Start up the server
----@return boolean
-function Server:startUp()
-    logger:info("Starting up...", self.protocol, self.hostname)
-    if not self:openModem() then
-        return false
-    end
-
-    rednet.host(self.protocol, self.hostname)
-
-    logger:info("%s server started!", self.protocol)
-
-    return true
-end
-
-
----Shut down the server
----@return boolean
-function Server:shutDown()
-    if not self.modemName then
-        return false
-    end
-
-    rednet.unhost(self.protocol)
-
-    self:closeModem()
-
-    return true
-end
-
-
 ---Handle a refresh request from a client
 ---@param clientId number
 ---@param data? table
 ---@return boolean
-function Server:handleRefresh(clientId, data)
+function StorageServer:handleRefresh(clientId, data)
     self.storageMap:populate(true)
     self.storageMap:save()
 
@@ -114,7 +76,7 @@ end
 ---@param clientId number
 ---@param data? table
 ---@return boolean, table?
-function Server:handleDataIoChests(clientId, data)
+function StorageServer:handleDataIoChests(clientId, data)
     local res = {
         inputChest = peripheral.getName(self.inputChest),
         outputChest = peripheral.getName(self.outputChest),
@@ -128,7 +90,7 @@ end
 ---@param clientId number
 ---@param data? table
 ---@return boolean, table
-function Server:handlePing(clientId, data)
+function StorageServer:handlePing(clientId, data)
     return true, {pong = true}
 end
 
@@ -137,7 +99,7 @@ end
 ---@param clientId number
 ---@param data? table
 ---@return boolean, table?
-function Server:handlePull(clientId, data)
+function StorageServer:handlePull(clientId, data)
     if not data then
         return false
     end
@@ -172,7 +134,7 @@ end
 ---@param clientId number
 ---@param data? table
 ---@return boolean, table?
-function Server:handlePush(clientId, data)
+function StorageServer:handlePush(clientId, data)
     if not data then
         return false
     end
@@ -200,73 +162,4 @@ function Server:handlePush(clientId, data)
     }
 
     return true, res
-end
-
-
----Listen for commands
----@return boolean
-function Server:_listen()
-    if not self:startUp() then
-        return false
-    end
-
-    while true do
-        local senderId, messageType, data = self:receiveData(nil, MessageType.CMD)
-        if not messageType then
-            goto continue
-        end
-        ---@cast data MessageCommandData
-
-        local commandType = data.type
-        local commandData = data.data
-
-        local handler = self.commandHandlers[commandType]
-
-        if not handler then
-            logger:warn("<%d|No handler for %s", senderId, messageType)
-            self:sendError(
-                senderId,
-                MessageErrorCode.UNKNOWN_COMMAND,
-                "No handler for %s",
-                messageType
-            )
-            goto continue
-        end
-
-        local handlerRes, handlerData = handler(self, senderId, commandData)
-
-        if not handlerRes then
-            logger:warn("<%d|Failed to handle %s", senderId, messageType)
-            self:sendError(
-                senderId,
-                MessageErrorCode.UNKNOWN,
-                "Failed to handle %s",
-                messageType
-            )
-            goto continue
-        else
-            self:sendData(senderId, MessageType.END, handlerData)
-        end
-
-        logger:info("<%d|Handled %s (%s)", senderId, messageType, commandType)
-
-        ::continue::
-    end
-end
-
-
----Listen for commands and safely shutdown the server on error/exit using xpcall
----@return boolean
-function Server:listen()
-    local status, listenRes = xpcall(self._listen, function (err)
-        logger:error("Error: %s", err)
-    end, self)
-
-    self:shutDown()
-
-    if not status or not listenRes then
-        return false
-    end
-
-    return true
 end
