+++ packages/library/lib-storage2/remote/examplePing.lua
@@ -1,5 +1,8 @@
 ---Example ping program
 package.path = package.path .. ";/usr/lib/?.lua"
-require("lib-storage2.remote.client.Client")
-local client = Client()
+
+require("lib-storage2.remote.StorageClient")
+
+local client = StorageClient()
+
 client:callCommand(client.ping)
